[xpu] adjust XPU_L3_LOCK_REQUIRED implements test=develop test=xpu
